fix: stat valeur can be float

+++ lib/data/db.dart
@@ -32,7 +32,7 @@ class DatabaseService {
           CREATE TABLE stat(
             id INTEGER PRIMARY KEY AUTOINCREMENT,
             name TEXT,
-            valeur INTEGER,
+            valeur FLOAT,
             persoId INTEGER
           )
         ''');
